Close picker on field blur
Currently, when we focus on the datepicker, the pickers
are shown. However, when we blur the field (that is, the
field loses focus), the picker will remain visible.

This makes it very unfriendly to keyboard users (among other
scenarios) because you have to hit escape or click on something
else on the page manually so that the pickers close.

By adding a listener to the blur event that calls the close method,
we ensure that the picker will automatically close when the field
loses focus.

This is related to https://github.com/chronotruck/vue-ctk-date-time-picker/issues/128